add comments about how the code works

+++ src/main/java/com/example/projectfilrouge/service/user/UserService.java
@@ -8,6 +8,7 @@ import com.example.projectfilrouge.entity.ConfirmationToken;
 import com.example.projectfilrouge.entity.Ticket;
 import com.example.projectfilrouge.entity.UserEntity;
 import com.example.projectfilrouge.entity.UserRole;
+import com.example.projectfilrouge.exception.NotAllowedException;
 import com.example.projectfilrouge.exception.NotFoundException;
 import com.example.projectfilrouge.repository.TicketRepository;
 import com.example.projectfilrouge.repository.UserRepository;
@@ -40,6 +41,8 @@ public class UserService implements UserDetailsService {
     private final EmailSender emailSender;
     private final ConfirmationTokenService confirmationTokenService;
 
+    //loadUserByUsername est une fonction à l'intérieur de UserDetailsService que tu surcharges pour donner
+    // l'info concernant l'endroit où on retrouve le userName
     @Override
     public UserDetails loadUserByUsername(String email)
             throws UsernameNotFoundException {
@@ -49,6 +52,7 @@ public class UserService implements UserDetailsService {
     }
 
     public void signUpUser(UserDto userDto) {
+        //A partir de ton userdto tu crée ton userentity
         UserEntity userEntity = new UserEntity(
                 userDto.getFirstName(),
                 userDto.getLastName(),
@@ -57,6 +61,7 @@ public class UserService implements UserDetailsService {
                 UserRole.USER
 
         );
+        //Verif si le user est présnet bdd ou non
         boolean userExists = userRepository.findByEmail(userEntity.getEmail())
                 .isPresent();
 
@@ -70,20 +75,26 @@ public class UserService implements UserDetailsService {
         String token = UUID.randomUUID()
                 .toString();
 
+       //Pour créer l'objet token
         ConfirmationToken confirmationToken = new ConfirmationToken(
                 token,
+                //La date et l'heure de maintenant
                 LocalDateTime.now(),
+                //Ici le timmer de 15 min
                 LocalDateTime.now()
                         .plusMinutes(15),
                 userEntity
         );
 
         userRepository.save(userEntity);
+        //Que tu insert en bdd ici
         confirmationTokenService.saveConfirmationToken(confirmationToken);
 
-        String link = "http://localhost:8080/api/users/registration/confirm?token=" + token;
+        // le lien de confirmation
+        String link = "http://localhost:8080/registration/confirm?token=" + token;
         emailSender.send(
                 userDto.getEmail(),
+                //C'est le texte en bas du mail
                 buildEmail(userDto.getFirstName(), link));
     }
 
@@ -91,21 +102,23 @@ public class UserService implements UserDetailsService {
     public String confirmToken(String token) {
         ConfirmationToken confirmationToken = confirmationTokenService
                 .getToken(token)
+                //Si jamais on a pas reçu le token
                 .orElseThrow(() ->
                         new IllegalStateException("token not found")
                 );
-
+//Si c'est déjà fait il te le dit
         if (confirmationToken.getConfirmedAt() != null) {
             throw new IllegalStateException("Email already confirmed");
         }
-
+//ça te donne l'heure d'expiration
         LocalDateTime expiredAt = confirmationToken.getExpiresAt();
-
+//Si c'est expiré ben mort
         if (expiredAt.isBefore(LocalDateTime.now())) {
             throw new IllegalStateException(("token expired"));
         }
-
+//Sinon ok on sauvegarde et en met à jour la date
         confirmationTokenService.setConfirmedAt(token);
+        //et on active le user
         userRepository.enableUser(confirmationToken.getUserEntity()
                 .getEmail());
         return "confirmed";
@@ -207,6 +220,8 @@ public class UserService implements UserDetailsService {
                 "</div></div>";
     }
 
+//C'est un findAll d'users que pour l'administrateur
+    //Afin de montrer qu'on peut restreindre la consultation de profils qu'à l'administrateur
     public ResponseEntity<List<UserEntity>> findAll() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         if(!auth.getAuthorities().contains(UserRole.ADMIN)) {
@@ -231,9 +246,11 @@ public class UserService implements UserDetailsService {
 
     public ResponseEntity<HttpStatus> updateUser(Long id, UserDto userDto) {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        //Condition pour que le bon user puisse faire le update
         if(!((UserEntity) auth.getPrincipal()).getId().equals(id)) {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
+        //On cherche par id et si il n'y est pas ben erreur
         UserEntity user =
                 userRepository.findById(id)
                         .orElseThrow(() -> new NotFoundException("user with id: " + id + "cannot be found"));
@@ -243,24 +260,40 @@ public class UserService implements UserDetailsService {
         user.setEmail(userDto.getEmail());
         user.setPassword(userDto.getPassword());
         userRepository.save(user);
+        // Response permet de changer le statut de la requête
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
-
+//Void ne renvoie rien fait juste un truc
     public void deleteUserById(Long id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        //Condition pour que le bon user puisse faire le update
+        if(!((UserEntity) auth.getPrincipal()).getId().equals(id)) {
+            throw new NotAllowedException();
+        }
         if ((userRepository.findById(id).isEmpty())) {
             throw new ResponseStatusException(HttpStatus.NO_CONTENT);
         }
         userRepository.deleteById(id);
     }
 
+//Permet d'obtenir l'historique de vente et d'achat de ticket
     public AllTicketDto getAllUserRelatedTicket() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         Long userId = ((UserEntity) auth.getPrincipal()).getId();
+        //On recupère la liste des tickets en vente par l'utilisateur lui même
         List<Ticket> ticketOnSale = ticketRepository.findTicketOnSale(userId);
+        //On récupère la liste d'achat des tickets par l'user lui même
         List<Ticket> ticketPurchased = ticketRepository.findTicketPurchased(userId);
+        //Créer un objet avec la liste de tous les tickets qui on un lien avec le user
         return new AllTicketDto(
+                //Objet qui contient une liste de ticket mis en vente et vendu
                 ticketOnSale.stream().map(ticket -> new TicketDto(ticket.getEventName(), ticket.getEventDate(), ticket.getPrice(), ticket.getDetails(), ticket.getState())).toList(),
+                //Objet qui contient une liste de ticket acheter
                 ticketPurchased.stream().map(ticket -> new TicketDto(ticket.getEventName(), ticket.getEventDate(), ticket.getPrice(), ticket.getDetails(), ticket.getState())).toList()
+                 //Stream est une biblio ou y a map
+                //Map permet de retouner une modification pour l'ensemble des tickets dans la liste (et de le renvoyer)
+                //ToList c'est tout ce qu'il a stocké en mémoire il en fait un eliste et me le renvoi
+                //Pour le mettre dans ticketOnSale ou ticketPurshed AllTicketDto
         );
 
     }
